+++ src/PublicRestApi/Courses/CourseRegistrationApi.php
@@ -6,16 +6,16 @@ namespace Dbp\CampusonlineApi\PublicRestApi\Courses;
 
 use Dbp\CampusonlineApi\PublicRestApi\AbstractApi;
 
-class LectureshipApi extends AbstractApi
+class CourseRegistrationApi extends AbstractApi
 {
     public const COURSE_UID_QUERY_PARAMETER_NAME = 'course_uid';
 
-    private const API_PATH = Common::API_PATH.'/lectureships';
+    private const API_PATH = Common::API_PATH.'/registrations';
 
     /**
      * @return iterable<LectureshipResource>
      */
-    public function getLectureships(array $queryParameters = [], array $options = []): iterable
+    public function getCourseRegistrations(array $queryParameters = [], array $options = []): iterable
     {
         return $this->getResources(self::API_PATH,
             LectureshipResource::class,
@@ -25,9 +25,9 @@ class LectureshipApi extends AbstractApi
     /**
      * @return iterable<LectureshipResource>
      */
-    public function getLectureshipsByCourseUid(string $courseUid, array $options = []): iterable
+    public function getCourseRegistrationsByCourseUid(string $courseUid, array $options = []): iterable
     {
-        return $this->getLectureships([
+        return $this->getCourseRegistrations([
             self::COURSE_UID_QUERY_PARAMETER_NAME => $courseUid,
         ], $options);
     }