+++ src/PublicRestApi/Rooms/RoomsApi.php
@@ -8,7 +8,7 @@ use Dbp\CampusonlineApi\PublicRestApi\AbstractApi;
 
 class RoomsApi extends AbstractApi
 {
-    private const ROOMS_API_PATH = Common::API_PATH.'/rooms';
+    private const API_PATH = Common::API_PATH.'/rooms';
 
     private const ROOM_UID_QUERY_PARAMETER_NAME = 'room_uid';
 
@@ -17,7 +17,7 @@ class RoomsApi extends AbstractApi
         $resource = $this->getResourceByIdentifierFromCollection(
             $identifier,
             self::ROOM_UID_QUERY_PARAMETER_NAME,
-            self::ROOMS_API_PATH,
+            self::API_PATH,
             RoomResource::class);
         assert($resource instanceof RoomResource);
 
@@ -31,7 +31,7 @@ class RoomsApi extends AbstractApi
         int $firstItemIndex = 0, int $maxNumItems = 30, array $options = []): iterable
     {
         return $this->getResourcesOffsetBased(
-            self::ROOMS_API_PATH, RoomResource::class,
+            self::API_PATH, RoomResource::class,
             $queryParameters, $firstItemIndex, $maxNumItems);
     }
 }